moved form in to component
