+++ src/modules/Checkout/Checkout.module.scss
@@ -2,22 +2,8 @@
   @extend %container;
 }
 
-.content,
-.form {
+.content {
   margin-top: 32px;
-}
-
-.form {
-  max-width: 360px;
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-
-  &__group {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-  }
 }
 
 .title {
@@ -26,6 +12,14 @@
   line-height: 32px;
   color: $main;
   margin-bottom: 12px;
+}
+
+.total {
+  font-size: 32px;
+  font-weight: 700;
+  line-height: 32px;
+  color: $main;
+  margin-top: 36px;
 }
 
 .products {
