+++ src/modules/ProductDetails/ProductDetails.module.scss
@@ -2,6 +2,25 @@
   @extend %container;
 }
 
+.empty {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+
+  &__img {
+    max-width: 420px;
+    width: 100%;
+  }
+
+  &__title {
+    color: $main;
+    font-size: 26px;
+    font-weight: 700;
+    line-height: 32px;
+    margin: 32px 0 16px;
+  }
+}
+
 .title {
   font-size: 32px;
   line-height: 41px;
